Moved CommandLookupKey to new file
